src/requirements.txt
matplotlib==3.9.2
networkx==3.3
numpy==1.24.3
openai==1.51.2
pandas==2.2.3
psycopg2==2.9.9
python-dotenv==1.0.1
rdflib==7.0.0
spacy==3.7.6
tiktoken==0.7.0
transformers==4.36.2
